pyproject.toml
[tool.poetry]
name = "ootd-back"
version = "0.1.0"
description = "OOTD Back-end"
authors = ["tejava317 <frogjj@kaist.ac.kr>"]
readme = "README.md"
repository = "https://github.com/tejava317/ootd-back"
license = "MIT"

packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
boto3 = "^1.35.92"
pytest = "^8.3.4"
httpx = "^0.28.1"
requests = "^2.32.3"
google-generativeai = "^0.8.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
